feat: implemented digital directional buttons for analog controls
Enabled the analog stick to simulate the behaviour of a digital directional pad. This is so that
functions that don't require analog control don't trigger unexpectedly when using the analog stick.
The digital directional pad mode of the analog stick only triggers a 'button press' when the analog
stick is extended over 95% of its range. An added benefit of this change is that button presses on
the analog stick don't linger when switching nunchuck modes. I have also applied this same technique
to regular C and Z button presses when calling Mouse and Keyboard methods, as these button presses
tend to linger as well, unless limiting them to when button state changes. An example of this is
when exiting the freehand mouse mode. The C button press no longer triggers a right mouse button
press when back in regular mouse mode.