refactor: replaced freehand mode with a joystick mode
After determining that a freehand mode is not feasible to implement, I converted this code to act as
a joystick/gamepad instead. Basic joystick features have been implemented, but some features are not
yet functional, such as the analog controls and the POV hat. Joystick data also isn't cleared when
entering and exiting the mode.