refactor: removed joystick and selection modes
The additional code was causing instability with the hardware, so these features were removed in
favour of more essential ones. The joystick mode offered an inferior user experience for gaming, so
this was an obvious mode to remove. The selection mode was only useful for choosing more than one
mode of the nunchuck. Since there was only the keyboard mode remaining, I decided to remove the
selection mode as well. The keyboard mode now gets activated during the mouse mode when the C button
is pressed while the nunchuck is tilted down.